+++ resources/js/humhub.cleanTheme.topNavigation.js
@@ -200,7 +200,7 @@ humhub.module('cleanTheme.topNavigation', function (module, require, $) {
      */
     const updateBtnStatus = function (btnId, moduleId, controllerId) {
         const state = humhub.modules.ui.view.getState();
-        const searchBtn = $('#' + btnId).parent();
+        const searchBtn = $('#' + btnId);
         if (searchBtn.length) {
             if (state.moduleId === moduleId && state.controllerId === controllerId) {
                 searchBtn.addClass('active');