+++ themes/Clean/views/humhub/layouts/main.php
@@ -9,6 +9,7 @@ use humhub\helpers\DeviceDetectorHelper;
 use humhub\helpers\Html;
 use humhub\modules\cleanTheme\assets\CleanThemeAsset;
 use humhub\modules\cleanTheme\assets\CleanThemeTopNavigationAsset;
+use humhub\modules\cleanTheme\models\Configuration;
 use humhub\modules\cleanTheme\Module;
 use humhub\modules\space\widgets\Chooser;
 use humhub\modules\user\widgets\AccountTopMenu;
@@ -19,19 +20,19 @@ use humhub\widgets\TopMenuRightStack;
 
 /** @var Module $module */
 $module = Yii::$app->getModule('clean-theme');
-$googleFontsCss2UrlParams = $module->configuration->getGoogleFontsCss2UrlParams();
+$googleFontsCss2UrlParams = $module?->configuration->getGoogleFontsCss2UrlParams();
 
 AppAsset::register($this);
 CleanThemeAsset::register($this);
 CleanThemeTopNavigationAsset::register($this);
 $this->registerJsConfig('cleanTheme.topNavigation', [
-    'hideTopMenuOnScrollDown' => $module->configuration->hideTopMenuOnScrollDown,
-    'hideBottomMenuOnScrollDown' => $module->configuration->hideBottomMenuOnScrollDown,
+    'hideTopMenuOnScrollDown' => $module?->configuration->hideTopMenuOnScrollDown ?? false,
+    'hideBottomMenuOnScrollDown' => $module?->configuration->hideBottomMenuOnScrollDown ?? false,
 ]);
 
 $bodyClasses = DeviceDetectorHelper::getBodyClasses();
 $bodyClasses[] = 'clean-theme';
-$bodyClasses[] = 'hh-ct-menu-style-' . $module->configuration->menuStyle;
+$bodyClasses[] = 'hh-ct-menu-style-' . ($module?->configuration->menuStyle ?? Configuration::MENU_STYLE_BACKGROUND);
 if (Yii::$app->user->isGuest) {
     $bodyClasses[] = 'hh-ct-is-guest';
 }
@@ -65,7 +66,7 @@ if (Yii::$app->user->isGuest) {
                 </div>
 
                 <ul id="top-menu-nav"
-                    class="nav<?= $module->configuration->hideTextInBottomMenuItems ? ' hide-menu-item-texts' : '' ?>">
+                    class="nav<?= $module?->configuration->hideTextInBottomMenuItems ? ' hide-menu-item-texts' : '' ?>">
                     <!-- load space chooser widget -->
                     <?= Chooser::widget() ?>